% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{aesfCRAWL}
\alias{aesfCRAWL}
\title{ARGOS Error Scale Factors}
\usage{
aesfCRAWL()
}
\value{
A dataframe with columns
\item{\code{lc}}{ARGOS location class}
\item{\code{lonESF}}{error scaling factor for longitude}
\item{\code{latESF}}{error scaling factor for latitude}
}
\description{
ARGOS Error Scale Factors for Location Classes
}
\details{
These are the multiplication factors used to scale location
accuracy for each location class that are used in \pkg{crawl}.
}
