% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{dcrw}
\alias{dcrw}
\title{Correlated Random Walk Filter}
\usage{
dcrw(data, subset = rep(TRUE, nrow(data)), tstep = 6/24, nu = 10,
  gamma = 0.5, span = 0.1, verbose = FALSE, optim = c("nlminb",
  "optim"), extrap = FALSE, parameters = NULL, esf = aesfCRAWL(),
  theta.zero = FALSE, common.tau = FALSE)
}
\arguments{
\item{data}{A dataframe representing the track (see details)}

\item{subset}{Logical vector indicating which rows of the data
frame be kept in the filtering}

\item{tstep}{The time step to predict to (in days)}

\item{nu}{The degrees of freedom parameter for the location errors}

\item{gamma}{The autocorrelation parameter used to estimate
initial parameters.}

\item{span}{The span parameter for the loess fits used to estimate
initial locations.}

\item{verbose}{Enable tracing information}

\item{optim}{The function used to minimize the negative log
likelihood.}

\item{extrap}{If TRUE, the final predicted state occurs
immediately before the last observation, otherwise the final
predicted state occurs immediately after the last observation.}

\item{parameters}{The TMB parameter list.}

\item{esf}{The error scale factors for the location classes.}

\item{theta.zero}{Should theta be fixed at zero.}

\item{common.tau}{Should a common tau parameter be fitted for
longitude and latitude.}
}
\value{
a list with components
  \item{\code{predicted}}{a dataframe of predicted locations}
  \item{\code{fitted}}{a dataframe of fitted locations}
  \item{\code{par}}{model parameter summary}
  \item{\code{data}}{the input dataframe}
  \item{\code{subset}}{the input subset vector}
  \item{\code{tstep}}{the prediction time step}
  \item{\code{theta.zero}}{has theta been fixed at zero}
  \item{\code{common.tau}}{has a common tau been fitted for lon and lat}
  \item{\code{opt}}{the object returned by the optimizer}
  \item{\code{tmb}}{the \pkg{TMB} object}
}
\description{
Correlated Random Walk Filter
}
\details{
Fit a correlated random walk to filter a track and predict
locations on a regular time step.

The input track is given as a dataframe where each row is an
observed location, with columns
\describe{
\item{'date'}{observation time (as GMT POSIXct),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude,}
\item{'lc'}{location class.}
}

The TMB parameter list can be specified directly with the
\code{parameters} argument. Otherwise suitable parameter values
are estimated from predicted locations estimated by fitting loess
smooths to the raw observations.

If \code{extrap} is \code{TRUE}, the last observations occur after
the last predicted location and the last fitted locations are
extrapolations, otherwise the final observations occur before the
final predicted locations and all fitted locations are
interpolated.

The filtering model assumes the errors in longitude and latitude
are proportional to scale factors determined by the location
class. The scale factors are speficied through the \code{aes}
argument. By default the function uses the same scaling factors
for location accuracy as used in \pkg{crawl} for ARGOS data.
}
