% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{dcrwModel}
\alias{dcrwModel}
\title{DCRW Movement Model}
\usage{
dcrwModel(fit)
}
\arguments{
\item{fit}{a fitted \code{dcrw} object}
}
\value{
A list containing
\item{\code{A}}{the transition matrix}
\item{\code{Q}}{the covariance matrix}
}
\description{
Generate Transition and Covariance Matrices for a DCRW model
}
\details{
Generates the transition and covariance matrices corresponding to
the dcrw state space model. Currently no polar adjustment is made.
}
