% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{interpolateTrack}
\alias{interpolateTrack}
\title{Interpolate Track}
\usage{
interpolateTrack(trk, tstep = 60 * 60)
}
\arguments{
\item{trk}{a dataframe representing a track}

\item{tstep}{the time step to reample to (in seconds)}
}
\value{
a dataframe with columns
  \item{\code{date}}{observation time (as POSIXct)}
  \item{\code{lon}}{interpolated longitude}
  \item{\code{lat}}{interpolated latitude}
}
\description{
Resample a Track by Linear Interpolation
}
\details{
Linearly interpolate in longitude and latitude to resample a track
back to a regular time step.

The input track is given as a dataframe where each row is an
observed location, with (at minimum) columns
\describe{
\item{'date'}{observation time (as GMT POSIXct),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude}
}
}
