% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\docType{package}
\name{Marseille-package}
\alias{Marseille-package}
\title{Track filtering and simulation}
\description{
Provides facilities for filtering and simulating tracks from
telemetry to support the Marseille RAATD workshop.
}
\author{
Ben Raymond, Ian Jonsen, Ryan Reisinger and Simon Wotherspoon
}
